+++ gulpfile.js/utilities/handle-errors.js
@@ -1,9 +1,9 @@
 // Define Dependencies
-var notify = require('gulp-notify')
+const notify = require('gulp-notify')
 
-// Handle Errors
+// Task
 module.exports = function () {
-  var args = Array.prototype.slice.call(arguments)
+  const args = Array.prototype.slice.call(arguments)
 
   // Send error to notification center with gulp-notify
   notify.onError({