+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.1.0 (December 28, 2019)
+
+- Update to Gulp 4
+- Update dependencies
+- Re-write Gulp tasks using new syntax (series, parallel, etc.)
+- Update README
+- Update Travis config file
+- Update tests
+- Add package-lock.json file
+- Rename `gulp/` directory to `gulpfile.js/`
+
 # 0.0.1 (January 30, 2016)
 
 - Initial release.